+++ bin/static/app.js
@@ -2344,4 +2344,4 @@
   };
 })();
 /* ui-bundle: f2b3a730fd0c60b7f82529bfdc699a0378f1581862d21576136480f73b00a0ba */
-/* ui-sources: 3f8d75136456819d5c5a4ec9e48c669696fc346368fc62a9e33ac6777f03dcab */
+/* ui-sources: 772ebcae0ec1efad9adee417c8ee4091e48a0bbcd5ddb1189f113ddd5814ff94 */